+++ Engine/Source/Runtime/Core/HAL/Memory/Memory.h
@@ -2,7 +2,7 @@
 #define CORE_LAH_MEMORY_H
 #include "../../InnerCore.h"
 
-void* RT_HAL_SYSAlloc(size_t uiSize);
-void RT_HAL_SYSFree(void* pMem, size_t uiSize);
+void* RT_HAL_SYSAlloc(u64 uiSize);
+void RT_HAL_SYSFree(void* pMem, u64 uiSize);
 
 #endif
