+++ Engine/Source/Runtime/Core/HAL/Memory/Src/MemoryWindows.cpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 #include <iostream>
 #include <Windows.h>
-void* RT_HAL_SYSAlloc(size_t uiSize)
+void* RT_HAL_SYSAlloc(u64 uiSize)
 {
     void* ptr = VirtualAlloc(nullptr, uiSize, MEM_COMMIT | MEM_RESERVE, PAGE_READWRITE);
     if (ptr == nullptr)
@@ -13,7 +13,7 @@ void* RT_HAL_SYSAlloc(size_t uiSize)
     return ptr;
 }
 
-void RT_HAL_SYSFree(void* pMem, size_t uiSize)
+void RT_HAL_SYSFree(void* pMem, u64 uiSize)
 {
     if (pMem != nullptr)
     {
